+++ Source/StarshatterWars/Display/CmdDlg.cpp
@@ -303,7 +303,7 @@ void UCmdDlg::OnExitClicked()
     if (Stars)
     {
         Mouse::Show(false);
-        Stars->SetGameMode((int)EMODE::MENU_MODE);
+        Stars->SetGameMode(EMODE::MENU_MODE);
     }
     else
     {
