+++ Source/StarshatterWars/Display/CmdIntelDlg.cpp
@@ -502,6 +502,6 @@ void UCmdIntelDlg::OnExitClicked()
     if (Stars)
     {
         Mouse::Show(false);
-        Stars->SetGameMode((int)EMODE::MENU_MODE);
+        Stars->SetGameMode(EMODE::MENU_MODE);
     }
 }
