+++ Source/StarshatterWars/Display/CmdOrdersDlg.cpp
@@ -213,6 +213,6 @@ void UCmdOrdersDlg::OnExitClicked()
     if (Stars)
     {
         Mouse::Show(false);
-        Stars->SetGameMode((int)EMODE::MENU_MODE);
+        Stars->SetGameMode(EMODE::MENU_MODE);
     }
 }
