+++ Source/StarshatterWars/Display/CmdTheaterDlg.cpp
@@ -186,7 +186,7 @@ void UCmdTheaterDlg::OnExitClicked()
     if (Stars)
     {
         Mouse::Show(false);
-        Stars->SetGameMode((int)EMODE::MENU_MODE);
+        Stars->SetGameMode(EMODE::MENU_MODE);
     }
 }
